Displayed Data for Polyp Bound Box Train Labels
Displayed in text Polyp Name for each Bound Box and the associated Bound Box coordinates using "xml.etree.ElementTree" library. So now we can use each of these labels with the particular image to train our DNN when we implement it.

Next I am going to draw the bound box dots onto an image for part of the data exploration, then next could be to do some further preprocessing on the images before putting them into training the DNN model.